Convert to single Entity class
All "entities" should extend this class. It implements location storage as well as all necessary internal movement.
