+++ Core/WumpusCore/WumpusCore.csproj
@@ -45,7 +45,6 @@
         <Compile Include="Controller\Controller.cs" />
         <Compile Include="Controller\ControllerState.cs" />
         <Compile Include="Entity\Entity.cs" />
-        <Compile Include="Entity\Mover.cs" />
         <Compile Include="Minigame.cs" />
         <Compile Include="LuckyCat\Cat.cs" />
         <Compile Include="MinigameController\MinigameController.cs" />
